implemented single responsibility principle
