+++ spec/tic_tac_toe_game/toggle_spec.rb
@@ -1,6 +1,5 @@
 require_relative '../../lib/tic_tac_toe_game/toggle.rb'
 
-
 RSpec.describe TicTacToeGame::Toggle do
     context "#other_turn" do
         it "Should switch from 'o' to 'x'" do
